feat: add Docker build context for AWS Lambda deployment
Add necessary folders and files to the Docker build context for deploying
SurrealDB as an AWS Lambda function. The added folders and files are
aws_lambda, src, lib, Cargo.toml, and build.rs. These files and folders were
previously ignored in the Docker build context. The changes were made to enable
easy deployment of SurrealDB as an AWS Lambda function.