refactor: update DynamoDB table schema to use new GSI1 hash key
This commit updates the DynamoDB table schema to use a new global secondary index (GSI1) with a hash key of "gsi1pk" and a range key of "pk". The previous hash key "bucket" and range key "key" are no longer used. This change was made to improve query performance and better match the data access patterns of the application. Additionally, the "key" attribute was removed as it was no longer needed.