feat(kvs): add support for specifying number of shards in DynamoDB path
This commit adds the ability to specify the number of shards in a DynamoDB path
by appending "?shards=N" to the end of the path. The regex pattern used to match
the DynamoDB path has been updated to include this optional parameter, and the
number of shards is extracted and passed to the Datastore constructor.

This change was done to support scaling out the DynamoDB backend for larger
workloads.