feat(kvs/dynamodb): add sharding support to distribute load evenly
This commit adds sharding support to distribute the load evenly across
multiple DynamoDB partitions. The `shards` field is added to the
`Datastore` struct and is passed to the `Transaction` struct when a new
transaction is started.

The `shard` function is added to generate a random shard index, and the
`bucket` function is removed as it is no longer needed. The `pk` field
in the `put_item` and `delete_item` methods is changed to `pk` from
`key`, and the `gsi1pk` field is added to GSI1 index. The query in the
`Query` method is updated to use `gsi1pk` instead of `bucket`.

WHY:
The sharding support is added to distribute the load evenly across
multiple DynamoDB partitions. This will help improve the performance of
the `Query`