feat: Update lambda Dockerfile to use Rust 1.70 and add target for aarch64-unknown-linux-gnu
The Dockerfile has been updated to use Rust 1.70 and install additional dependencies required to build for the aarch64-unknown-linux-gnu target. A new target for aarch64-unknown-linux-gnu has been added to Rustup. The build command has been updated to use the new target and install the binary for the surrealdb application to the target directory.